% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyData.bim.R
\docType{data}
\name{DummyData.bim}
\alias{DummyData.bim}
\title{PLINK .bim file}
\format{
This follows PLINK general format
\describe{
\item{Column 1}{Chromosome ID}
\item{Column 2}{SNP ID}
\item{Column 3}{Position of centimorgans}
\item{Column 4}{Base-pair coordinate}
\item{Column 5}{Minor Allele}
\item{Column 6}{Reference Allele}
}
}
\usage{
DummyData.bim
}
\description{
PLINK .bim file
}
\keyword{datasets}
