% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyData.fam.R
\docType{data}
\name{DummyData.fam}
\alias{DummyData.fam}
\title{PLINK .fam file}
\format{
This follows PLINK general format
\describe{
\item{Column 1}{Family ID}
\item{Column 2}{Individual ID}
\item{Column 3}{Father's ID}
\item{Column 4}{Mother's ID}
\item{Column 5}{Sex}
\item{Column 6}{Phenotype value}
}
}
\usage{
DummyData.fam
}
\description{
PLINK .fam file
}
\keyword{datasets}
