% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_quantitative.R
\name{GWAS_quantitative}
\alias{GWAS_quantitative}
\title{GWAS_quantitative function
This function performs GWAS using plink2 and outputs the GWAS summary statistics file with additive SNP effects named Q_out.trd.sum}
\usage{
GWAS_quantitative(
  plink_path,
  b_file,
  Qphe_discovery,
  Qcov_discovery,
  thread = 20,
  summary_output = "Q_out.trd.sum"
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{Qphe_discovery}{Name (with file extension) of the phenotype file containing family ID, individual ID and phenotype of the discovery dataset as columns, without heading}

\item{Qcov_discovery}{Name (with file extension) of the covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the discovery dataset as columns, without heading}

\item{thread}{Number of threads used}

\item{summary_output}{Name of the SNP effects of the GWAS summary statistics file specified by the user}
}
\value{
This function will perform GWAS and output
\item{Q_out.trd.sum}{GWAS summary statistics file with additive SNP effects}
}
\description{
GWAS_quantitative function
This function performs GWAS using plink2 and outputs the GWAS summary statistics file with additive SNP effects named Q_out.trd.sum
}
\examples{
\dontrun{
x <- GWAS_quantitative(plink_path, DummyData, Qphe_discovery, Qcov_discovery, 
                       thread = 20, summary_output = "Q_out.trd.sum")
head(x) #to read the head of all columns in Q_out.trd.sum file
x$V1 #to extract the chromosome number (CHROM)
x$V2 #to extract the base pair position (POS)
x$V3 #to extract the SNP ID (ID)
x$V4 #to extract the reference allele (REF)
x$V5 #to extract the alternate allele (ALT)
x$V6 #to extract the minor allele (A1)
x$V7 #to extract the type of test performed (TEST)
x$V8 #to extract the nmber of allele observations (OBS_CT)
x$V9 #to extract the SNP effect (BETA)
x$V10 #to extract the standard error of each SNP effect (SE)
x$V11 #to extract the test statistic (T_STAT)
x$V12 #to extract the p value (P)
x$V13 #to extract the error code (ERRCODE)
}
}
\keyword{gwas}
