% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qcov_discovery.R
\docType{data}
\name{Qcov_discovery}
\alias{Qcov_discovery}
\title{Covariate data file of the discovery dataset when the outcome is quantitative
This contains covariate information of the individuals in the discovery dataset following confounders}
\format{
A dataframe with 6426 rows and 18 columns
\describe{
\item{Column 1}{Family ID}
\item{Column 2}{Individual ID}
\item{Column 3}{Standardized covariate}
\item{Column 4}{Square of the standardized covariate}
\item{Column 5}{Confounder 1}
\item{Column 6}{Confounder 2}
\item{Column 7}{Confounder 3}
\item{Column 8}{Confounder 4}
\item{Column 9}{Confounder 5}
\item{Column 10}{Confounder 6}
\item{Column 11}{Confounder 7}
\item{Column 12}{Confounder 8}
\item{Column 13}{Confounder 9}
\item{Column 14}{Confounder 10}
\item{Column 15}{Confounder 11}
\item{Column 16}{Confounder 12}
\item{Column 17}{Confounder 13}
\item{Column 18}{Confounder 14}
}
}
\usage{
Qcov_discovery
}
\description{
Covariate data file of the discovery dataset when the outcome is quantitative
This contains covariate information of the individuals in the discovery dataset following confounders
}
\keyword{datasets}
