% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qphe_discovery.R
\docType{data}
\name{Qphe_discovery}
\alias{Qphe_discovery}
\title{Phenotype data file of the discovery dataset when the outcome is quantitative.
This contains phenotype information of the individuals in the discovery dataset.}
\format{
A dataframe with 800 rows and 3 columns
\describe{
\item{Column 1}{Family ID}
\item{Column 2}{Individual ID}
\item{Column 3}{Phenotype}
}
}
\usage{
Qphe_discovery
}
\description{
Phenotype data file of the discovery dataset when the outcome is quantitative.
This contains phenotype information of the individuals in the discovery dataset.
}
\keyword{datasets}
