% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qphe_target.R
\docType{data}
\name{Qphe_target}
\alias{Qphe_target}
\title{Phenotype data file of the target dataset when the outcome is quantitative.
This contains phenotype information of the individuals in the target dataset.}
\format{
A dataframe with 200 rows and 3 columns
\describe{
\item{Column 1}{Family ID}
\item{Column 2}{Individual ID}
\item{Column 3}{Phenotype}
}
}
\usage{
Qphe_target
}
\description{
Phenotype data file of the target dataset when the outcome is quantitative.
This contains phenotype information of the individuals in the target dataset.
}
\keyword{datasets}
