% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_permuted_binary.R
\name{summary_permuted_binary}
\alias{summary_permuted_binary}
\title{summary_permuted_binary function
This function uses plink2 and outputs the p value of permuted model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum}
\usage{
summary_permuted_binary(
  Bphe_target,
  Bcov_target,
  iterations = 1000,
  add_score = "B_add.sscore",
  gxe_score = "B_gxe.sscore"
)
}
\arguments{
\item{Bphe_target}{Phenotype file containing family ID, individual ID and phenotype of the target dataset as columns, without heading}

\item{Bcov_target}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the target dataset as columns, without heading}

\item{iterations}{Number of iterations used in permutation}

\item{add_score}{The .sscore file generated using additive SNP effects of GWEIS summary statistics}

\item{gxe_score}{The .sscore file generated using interaction SNP effects of GWEIS summary statistics}
}
\value{
This function will output
\item{B_permuted_p.txt}{the p value of the permuted model}
}
\description{
summary_permuted_binary function
This function uses plink2 and outputs the p value of permuted model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum
}
\examples{
\dontrun{ 
x <- summary_permuted_binary(Bphe_target, Bcov_target)
x
}
}
\keyword{regression}
\keyword{summary}
