% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_permuted_quantitative.R
\name{summary_permuted_quantitative}
\alias{summary_permuted_quantitative}
\title{summary_permuted_quantitative function
This function uses plink2 and outputs the p value of permuted model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named Q_trd.sum, Q_add.sum and Q_gxe.sum}
\usage{
summary_permuted_quantitative(
  Qphe_target,
  Qcov_target,
  iterations = 1000,
  add_score = "Q_add.sscore",
  gxe_score = "Q_gxe.sscore"
)
}
\arguments{
\item{Qphe_target}{Phenotype file containing family ID, individual ID and phenotype of the target dataset as columns, without heading}

\item{Qcov_target}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the target dataset as columns, without heading}

\item{iterations}{Number of iterations used in permutation}

\item{add_score}{The .sscore file generated using additive SNP effects of GWEIS summary statistics}

\item{gxe_score}{The .sscore file generated using interaction SNP effects of GWEIS summary statistics}
}
\value{
This function will output
\item{Q_permuted_p.txt}{the p value of the permuted model}
}
\description{
summary_permuted_quantitative function
This function uses plink2 and outputs the p value of permuted model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named Q_trd.sum, Q_add.sum and Q_gxe.sum
}
\examples{
\dontrun{ 
x <- summary_permuted_quantitative(Qphe_target, Qcov_target)
x 
}
}
\keyword{regression}
\keyword{summary}
