% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_binary.R
\name{PRS_binary}
\alias{PRS_binary}
\title{PRS_binary function
This function uses plink2 and outputs Polygenic Risk Scores (PRSs) of all the individuals, using pre-generated GWAS and/or GWEIS summary statistics. Note that the input files used in this function can be generated by using GWAS_binary and/or GWEIS_binary functions. It is recommended to save the output in a user-specified file (see examples).}
\usage{
PRS_binary(plink_path, b_file, summary_input)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{summary_input}{Name of the summary statistics file specified by the user}
}
\value{
This function will output
\item{B_trd.sscore}{PRSs for each individual}
}
\description{
PRS_binary function
This function uses plink2 and outputs Polygenic Risk Scores (PRSs) of all the individuals, using pre-generated GWAS and/or GWEIS summary statistics. Note that the input files used in this function can be generated by using GWAS_binary and/or GWEIS_binary functions. It is recommended to save the output in a user-specified file (see examples).
}
\examples{
\dontrun{ 
x <- PRS_binary(plink_path, DummyData, summary_input = "B_out.trd.sum")
sink("B_trd.sscore") #to create a file in the working directory
write.table(x, sep = " ", row.names = FALSE, quote = FALSE) #to write the output
sink() #to save the output
head(x) #to read the head of all columns in the output
x$V1 #to extract the family ID's of target dataset (FID)
x$V2 #to extract the individual ID's of target dataset (IID)
x$V3 #to extract the number of alleles across scored variants (ALLELE_CT)
x$V4 #to extract the sum of named allele dosages (NAMED_ALLELE_DOSAGE_SUM)
x$V5 #to extract the polygenic risk scores (SCORE1_AVG)
y <- PRS_binary(plink_path, DummyData, summary_input = "B_out.add.sum")
sink("B_add.sscore") #to create a file in the working directory
write.table(y, sep = " ", row.names = FALSE, quote = FALSE) #to write the output
sink() #to save the output
z <- PRS_binary(plink_path, DummyData, summary_input = "B_out.gxe.sum")
sink("B_gxe.sscore") #to create a file in the working directory
write.table(z, sep = " ", row.names = FALSE, quote = FALSE) #to write the output
sink() #to save the output
}
}
\keyword{profile}
\keyword{prs}
\keyword{scores}
