% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_quantitative.R
\name{PRS_quantitative}
\alias{PRS_quantitative}
\title{PRS_binary function
This function uses plink2 and outputs PRSs of each individual in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named Q_out.trd.sum, Q_out.add.sum or Q_out.gxe.sum}
\usage{
PRS_quantitative(
  plink_path,
  b_file,
  summary_input = "Q_out.trd.sum",
  summary_output = "Q_trd"
)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application}

\item{b_file}{Prefix of the binary files, where all .fam, .bed and .bim files have a common prefix}

\item{summary_input}{Name of the summary statistics file specified by the user}

\item{summary_output}{Name of the PRS file generated using provided summary statistics file specified by the user}
}
\value{
This function will output
\item{Q_trd.sscore}{PRSs for each target individual using GWAS additive effects}
}
\description{
PRS_binary function
This function uses plink2 and outputs PRSs of each individual in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named Q_out.trd.sum, Q_out.add.sum or Q_out.gxe.sum
}
\examples{
\dontrun{ 
x <- PRS_quantitative(plink_path, DummyData, summary_input = "Q_out.trd.sum", 
                      summary_output = "Q_trd")
head(x[[1]]) #to read the head of all columns in Q_trd.sscore file
x$V1 #to extract the family ID's of target dataset (FID)
x$V2 #to extract the individual ID's of target dataset (IID)
x$V3 #to extract the number of alleles across scored variants (ALLELE_CT)
x$V4 #to extract the sum of named allele dosages (NAMED_ALLELE_DOSAGE_SUM)
x$V5 #to extract the polygenic risk scores (SCORE1_AVG)
}
}
\keyword{profile}
\keyword{prs,}
\keyword{scores}
