% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_regular_binary.R
\name{summary_regular_binary}
\alias{summary_regular_binary}
\title{summary_regular_binary function
This function uses plink2 and outputs the summary of regular model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum}
\usage{
summary_regular_binary(
  Bphe_target,
  Bcov_target,
  trd_score = "B_trd.sscore",
  add_score = "B_add.sscore",
  gxe_score = "B_gxe.sscore",
  Model,
  summary_output = "Bsummary.txt",
  risk_output = "Individual_risk_values.txt"
)
}
\arguments{
\item{Bphe_target}{Phenotype file containing family ID, individual ID and phenotype of the target dataset as columns, without heading}

\item{Bcov_target}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the target dataset as columns, without heading}

\item{trd_score}{The .sscore file generated using additive SNP effects of GWAS summary statistics}

\item{add_score}{The .sscore file generated using additive SNP effects of GWEIS summary statistics}

\item{gxe_score}{The .sscore file generated using interaction SNP effects of GWEIS summary statistics}

\item{Model}{Specify the model number (1: y = PRS_trd + E + PRS_trd x E + confounders, 2: y = PRS_add + E + PRS_add x E + confounders, 3: y = PRS_add + E + PRS_gxe x E + confounders, 4: y = PRS_add + E + PRS_gxe + PRS_gxe x E + confounders, 5: y = PRS_add + E + E^2 + PRS_gxe + PRS_gxe x E + confounders, where y is the outcome variable, E is the covariate of interest, PRS_trd and PRS_add are the polygenic risk scores computed using additive SNP effects of GWAS and GWEIS summary statistics respectively, and PRS_gxe is the polygenic risk scores computed using GxE interaction SNP effects of GWEIS summary statistics.)}

\item{summary_output}{Name of the model summary file specified by the user}

\item{risk_output}{Name of the file containing risk scores of the target individuals specified by the user}
}
\value{
This function will output
\item{Bsummary.txt}{the summary of the fitted model}
\item{Individual_risk_values.txt}{the estimated risk values of individuals in the target sample}
}
\description{
summary_regular_binary function
This function uses plink2 and outputs the summary of regular model in the target dataset, using pre-generated GWAS and GWEIS summary statistics files named B_trd.sum, B_add.sum and B_gxe.sum
}
\examples{
\dontrun{ 
x <- summary_regular_binary(Bphe_target, Bcov_target, 
                            add_score = "B_add.sscore", 
                            gxe_score = "B_gxe.sscore", 
                            Model = 5, 
                            summary_output = "Bsummary.txt", 
                            risk_output = "Individual_risk_values.txt") 
x[[1]][[1]] #to obtain the content of the Bsummary.txt file
x[[1]][[2]] #to extract "Call" of the model summary
x[[1]][[3]] #to extract terms of the model summary
x[[1]][[4]] #to extract family information of the model summary
x[[1]][[5]] #to extract deviance information of the model summary
x[[1]][[6]] #to extract AIC information of the model summary
x[[1]][[7]] #to extract contrasts of the model summary
x[[1]][[8]] #to extract degrees of freedom (df) of residuals of 
            #the model summary
x[[1]][[9]] #to extract "Null deviance" of the model summary
x[[1]][[10]] #to extract degrees of freedom (df) of null deviance 
             #of the model summary
x[[1]][[11]] #to extract "iter" of the model summary
x[[1]][[12]] #to extract deviance residuals 
x[[1]][[13]] #to extract regrerssion coefficients of the model summary
x[[1]][[14]] #to extract aliesed information of the model summary
x[[1]][[15]] #to extract dispersion information of the model summary
x[[1]][[16]] #to extract degrees of freedom of the model summary
x[[1]][[17]] #to extract unscaled variance covariance matrix of all variables
x[[1]][[18]] #to extract scaled variance covariance matrix of all variables
head(x[[2]]) #to read the head of all columns in Individual_risk_values.txt file
x[[2]][,1] #to extract the column containing family ID's  
x[[2]][,2] #to extract the column containing individual ID's
x[[2]][,3] #to extract the column containing predicted risk scores 
}
}
\keyword{regression}
\keyword{risk}
\keyword{scores}
\keyword{summary}
