% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_regular_quantitative.R
\name{summary_regular_quantitative}
\alias{summary_regular_quantitative}
\title{summary_regular_quantitative function
This function outputs the summary of regular model and final risk score values of each individual in the target dataset using pre-generated Polygenic Risk Scores (PRSs) of all the individuals. Note that the input used in this function can be generated by using PRS_quantitative function.}
\usage{
summary_regular_quantitative(
  Qphe_target,
  Qcov_target,
  trd_score = NULL,
  add_score = NULL,
  gxe_score = NULL,
  Model
)
}
\arguments{
\item{Qphe_target}{Phenotype file containing family ID, individual ID and phenotype of the target dataset as columns, without heading}

\item{Qcov_target}{Covariate file containing family ID, individual ID, standardized covariate, square of standardized covariate, and/or confounders of the target dataset as columns, without heading}

\item{trd_score}{PRSs generated using additive SNP effects of GWAS summary statistics}

\item{add_score}{PRSs generated using additive SNP effects of GWEIS summary statistics}

\item{gxe_score}{PRSs generated using interaction SNP effects of GWEIS summary statistics}

\item{Model}{Specify the model number (1: y = PRS_trd + E + PRS_trd x E + confounders, 2: y = PRS_add + E + PRS_add x E + confounders, 3: y = PRS_add + E + PRS_gxe x E + confounders, 4: y = PRS_add + E + PRS_gxe + PRS_gxe x E + confounders, where y is the outcome variable, E is the covariate of interest, PRS_trd and PRS_add are the polygenic risk scores computed using additive SNP effects of GWAS and GWEIS summary statistics respectively, and PRS_gxe is the polygenic risk scores computed using GxE interaction SNP effects of GWEIS summary statistics.)}
}
\value{
This function will output
\item{Qsummary.txt}{the summary of the fitted model}
\item{Individual_risk_values.txt}{the estimated risk values of individuals in the target sample}
}
\description{
summary_regular_quantitative function
This function outputs the summary of regular model and final risk score values of each individual in the target dataset using pre-generated Polygenic Risk Scores (PRSs) of all the individuals. Note that the input used in this function can be generated by using PRS_quantitative function.
}
\examples{
\dontrun{
a <- GWAS_quantitative(plink_path, DummyData, Qphe_discovery, Qcov_discovery)
b <- GWEIS_quantitative(plink_path, DummyData, Qphe_discovery, Qcov_discovery)
p <- PRS_quantitative(plink_path, DummyData, summary_input = a)
q <- PRS_quantitative(plink_path, DummyData, summary_input = b[[1]])
r <- PRS_quantitative(plink_path, DummyData, summary_input = b[[2]])
summary_regular_quantitative(Qphe_target, Qcov_target, 
                            trd_score = p,
                            Model = 1)
summary_regular_quantitative(Qphe_target, Qcov_target, 
                            add_score = q,
                            Model = 2)
summary_regular_quantitative(Qphe_target, Qcov_target, 
                            add_score = q, 
                            gxe_score = r, 
                            Model = 3) 
x <- summary_regular_quantitative(Qphe_target, Qcov_target, 
                            add_score = q, 
                            gxe_score = r, 
                            Model = 4) 
sink("Qsummary.txt") #to create a file in the working directory
print(x[[1]][[1]]) #to write the output
sink() #to save the output
sink("Individual_risk_values.txt") #to create a file in the working directory
write.table(x[[2]], sep = " ", row.names = FALSE, col.names = FALSE, 
quote = FALSE) #to write the output
sink() #to save the output
x[[1]][[1]] #to obtain the model summary output
x[[1]][[2]] #to extract "Call" of the model summary
x[[1]][[3]] #to extract terms of the model summary
x[[1]][[4]] #to extract the residuals
x[[1]][[5]] #to extract regrerssion coefficients of the model summary
x[[1]][[6]] #to extract aliesed information of the model summary
x[[1]][[7]] #to extract "sigma" (residual standard error) information 
            #of the model summary
x[[1]][[8]] #to extract degrees of freedom of the model summary
x[[1]][[9]] #to extract the R squared value of the model summary
x[[1]][[10]] #to extract the adjusted R squared value of the model summary
x[[1]][[11]] #to extract the test statistic values of the model summary
x[[1]][[12]] #to extract unscaled variance covariance matrix of all variables
head(x[[2]]) #to view the head of the predicted risk values of target individuals
x[[2]][,1] #to extract the column containing family ID's 
x[[2]][,2] #to extract the column containing individual ID's 
x[[2]][,3] #to extract the column containing predicted risk scores 
}
}
\keyword{regression}
\keyword{risk}
\keyword{scores}
\keyword{summary}
